pyproject.toml
[project]
name = "ecore_automation"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "playwright==1.48.0",
    "pytest-playwright==0.5.2",
]

[project.optional-dependencies]
dev = [
    "pip-tools==7.4.1",
    "ruff==0.7.4",
    "mypy==1.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",              
    "--maxfail=2",  # stop after 2 failures   
    # "--headed", # open browser durring tests 
    # "-v",           # verbose output      
]
markers = [
    "e2e: end-to-end tests",
    "only: run specific tests"
]

[tools.playwright]
headless = true
browser = "chromium"
viewport = { width = 1920, height = 1080 }
screenshot = "only-on-failure"
video = "retain-on-failure"
trace = "retain-on-failure"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
